moved i18n files; removed unnecessary files
