+++ scratch-blocks-develop/i18n/create_msg.js
