+++ scratch-blocks-develop/i18n/strings_to_json_mapping.json
